Created a background js and updated manifest.json
Created background.js , thus required to register it to manifest.json and inform while file to reference. It is a non-persistent script , so runtime.onInstalled is need for listening event and inside it contains a storage API (the persistent part). It will set value using that API but including API will require register again on the manifest. Thus manifest uses permission to register the "Storage" API.